Update windows build script and make it build both 32-bit and 64-bit via command line flag.

+++ build_runtime_win.pl
@@ -3,8 +3,20 @@ use File::Spec;
 use File::Basename;
 use File::Copy;
 use File::Path;
+use Getopt::Long;
 my $root = File::Spec->rel2abs( dirname($0) );
 my $monoroot = File::Spec->rel2abs( dirname($0) . "/../mono" );
+
+my $skipbuild=0;
+my $debug = 0;
+my $build64 = 0;
+
+GetOptions(
+   "skipbuild=i"=>\$skipbuild,
+   "debug=i"=>\$debug,
+   "build64=i"=>\$build64,
+) or die ("illegal cmdline options");
+
 if ($ENV{UNITY_THISISABUILDMACHINE})
 {
 	print "rmtree-ing $root/builds because we're on a buildserver, and want to make sure we don't include old artifacts\n";
@@ -13,22 +25,45 @@ if ($ENV{UNITY_THISISABUILDMACHINE})
 	print "not rmtree-ing $root/builds, as we're not on a buildmachine";
 }
 
+my $config = "Release";
+my $platform = "Win32";
+my $embedDir = "win32";
+my $binDirectory = "bin";
 
-CompileVCProj("$monoroot/msvc/mono.sln","Release",0);
-my $remove = "$root/builds/embedruntimes/win32/libmono.bsc";
-if (-e $remove)
+if ($debug)
 {
-	unlink($remove) or die("can't delete libmono.bsc");
+	$config = "Debug";
 }
 
-mkpath("$root/builds/embedruntimes/win32");
-mkpath("$root/builds/monodistribution/bin");
-copy("$monoroot/msvc/Win32/bin/mono-2.0.dll","$root/builds/embedruntimes/win32/mono.dll");
-copy("$monoroot/msvc/Win32/bin/mono-2.0.pdb","$root/builds/embedruntimes/win32/mono.pdb");
-copy("$monoroot/msvc/Win32/bin/mono.exe","$root/builds/monodistribution/bin/mono.exe");
-copy("$monoroot/msvc/Win32/bin/mono-2.0.dll","$root/builds/monodistribution/bin/");
-copy("$monoroot/msvc/Win32/bin/mono-2.0.pdb","$root/builds/monodistribution/bin/");
+if ($build64)
+{
+	$platform = "x64";
+	$embedDir = "win64";
+	$binDirectory = "bin-x64";
+}
 
+if (not $skipbuild)
+{
+	CompileVCProj("$monoroot/msvc/mono.sln","$config|$platform",0);
+	my $remove = "$root/builds/embedruntimes/$embedDir/libmono.bsc";
+	if (-e $remove)
+	{
+		unlink($remove) or die("can't delete libmono.bsc");
+	}
+}
+
+mkpath("$root/builds/embedruntimes/$embedDir");
+mkpath("$root/builds/monodistribution/$binDirectory");
+copy("$monoroot/msvc/$platform/bin/mono-2.0.dll","$root/builds/embedruntimes/$embedDir/mono.dll");
+copy("$monoroot/msvc/$platform/bin/mono-2.0.pdb","$root/builds/embedruntimes/$embedDir/mono.pdb");
+copy("$monoroot/msvc/$platform/bin/mono.exe","$root/builds/monodistribution/$binDirectory/mono.exe");
+copy("$monoroot/msvc/$platform/bin/mono-2.0.dll","$root/builds/monodistribution/$binDirectory/");
+copy("$monoroot/msvc/$platform/bin/mono-2.0.pdb","$root/builds/monodistribution/$binDirectory/");
+
+if ($ENV{UNITY_THISISABUILDMACHINE})
+{
+	system("echo mono-runtime-$embedDir = $ENV{'BUILD_VCS_NUMBER'} > $monoroot\\builds\\versions.txt");
+}
 
 sub CompileVCProj
 {
